# commons/src/main/java/com/ordermanagementsystem/dto/PasswordChangeDTO.java
package com.ordermanagementsystem.dto;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class PasswordChangeDTO {
    @NotNull
    @Size(min = 8)
    private String password;
}